Move SettingsStore methods to SettingsStoreMethods to use them independently within Yosemite
